backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
anthropic==0.39.0
openai==1.54.0
aiofiles==24.1.0
python-dotenv==1.0.1
google-generativeai==0.8.3
zhipuai
